# swagger/open_api.yml
{"openapi":"3.0.1","info":{"title":"Product info","description":"Test project for Setronica","version":"1.0.0"},"servers":[{"url":"http://localhost:8099","description":"Generated server url"}],"tags":[{"name":"Admin controller","description":"the controller contains operations for adding, changing, deleting products"},{"name":"User controller","description":"intended for receiving products by customers"}],"paths":{"/product":{"get":{"tags":["Admin controller"],"summary":"Get product by ID","description":"allows to get a product by a unique identifier","operationId":"getProductById_1","parameters":[{"name":"id","in":"query","description":"unique product identifier","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"product id not specified","content":{"application/json":{"schema":{"type":"string"}}}},"404":{"description":"product not found","content":{"application/json":{"schema":{"type":"string"}}}},"200":{"description":"Successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}},"put":{"tags":["Admin controller"],"summary":"Update","description":"allows update product","operationId":"updateProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"product successfully updated","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}},"500":{"description":"an error occurred on the server while updating product","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"one or more of the required fields of the product is not filled","content":{"application/json":{"schema":{"type":"string"}}}}}},"post":{"tags":["Admin controller"],"summary":"New product","description":"allows to create a new product","operationId":"postProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"201":{"description":"product successfully created","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}},"400":{"description":"one or more of the required fields of the product is not filled","content":{"application/json":{"schema":{"type":"string"}}}},"500":{"description":"an error occurred on the server while creating a new product","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"tags":["Admin controller"],"summary":"Delete","description":"allows delete product","operationId":"deleteProduct","parameters":[{"name":"id","in":"query","description":"unique product identifier","required":true,"schema":{"minimum":1,"type":"integer","format":"int64"}}],"responses":{"200":{"description":"product successfully deleted","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}},"404":{"description":"product not found","content":{"application/json":{"schema":{"type":"string"}}}},"500":{"description":"an error occurred on the server while deleting product","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"an error occurred on the server while deleting product","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/product_resource/title":{"get":{"tags":["User controller"],"summary":"Get products by Title","description":"allows to get a products by part of title, language and currency","operationId":"getProductByTitle","parameters":[{"name":"title","in":"query","description":"full title or part of it","required":true,"schema":{"type":"string"}},{"name":"language","in":"query","description":"product data language","required":true,"schema":{"type":"string"}},{"name":"currency","in":"query","description":"product currency of price","required":true,"schema":{"type":"string"}},{"name":"size","in":"query","description":"number of results per page","required":true,"schema":{"minimum":0,"type":"integer","format":"int32"}},{"name":"page","in":"query","description":"number of page","required":true,"schema":{"minimum":0,"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}},"404":{"description":"products with this title, language, currency not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}},"400":{"description":"one of the required fields is not filled","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}}}}},"/product_resource/id":{"get":{"tags":["User controller"],"summary":"Get product by ID","description":"allows to get a product by a unique identifier, language and currency","operationId":"getProductById","parameters":[{"name":"id","in":"query","description":"unique product identifier","required":true,"schema":{"minimum":1,"type":"integer","format":"int64"}},{"name":"language","in":"query","description":"product data language","required":true,"schema":{"type":"string"}},{"name":"currency","in":"query","description":"product currency of price","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"product not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}},"200":{"description":"Successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}},"400":{"description":"product id not specified","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}}}}},"/product_resource/description":{"get":{"tags":["User controller"],"summary":"Get products by description","description":"allows to get a products by part of description, language and currency","operationId":"getProductByPartOfDescription","parameters":[{"name":"description","in":"query","description":"full description or part of it","required":true,"schema":{"type":"string"}},{"name":"language","in":"query","description":"product data language","required":true,"schema":{"type":"string"}},{"name":"currency","in":"query","description":"product currency of price","required":true,"schema":{"type":"string"}},{"name":"size","in":"query","description":"number of results per page","required":true,"schema":{"minimum":0,"type":"integer","format":"int32"}},{"name":"page","in":"query","description":"number of page","required":true,"schema":{"minimum":0,"type":"integer","format":"int32"}}],"responses":{"404":{"description":"products with this description, language, currency not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}},"200":{"description":"Successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}},"400":{"description":"one of the required fields is not filled","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}}}}},"/product_resource/all":{"get":{"tags":["User controller"],"summary":"Get all products by description","description":"allows to get all products with data in the specified language and currency","operationId":"getAllProducts","parameters":[{"name":"language","in":"query","description":"product data language","required":true,"schema":{"type":"string"}},{"name":"currency","in":"query","description":"product currency of price","required":true,"schema":{"type":"string"}},{"name":"size","in":"query","description":"number of results per page","required":true,"schema":{"minimum":0,"type":"integer","format":"int32"}},{"name":"page","in":"query","description":"number of page","required":true,"schema":{"minimum":0,"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}},"404":{"description":"products with data in the specified language and currency not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}},"400":{"description":"one of the required fields is not filled","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exception"}}}}}}}},"components":{"schemas":{"Exception":{"properties":{"errorCode":{"type":"integer","format":"int32","example":404},"errorMessage":{"type":"string","example":"not found"}}},"Language":{"type":"object","properties":{"languageId":{"type":"string"}},"description":"product data language","example":"en"},"PriceCurrency":{"type":"object","properties":{"currencyId":{"type":"string"}},"description":"product currency of price","example":"usd"},"Product":{"required":["title"],"type":"object","properties":{"productId":{"type":"integer","description":"required field when searching by id or updating, deleting product, but not necessary when creating a new product or searching in other fields","format":"int64","example":1},"title":{"type":"string","description":"required field when creating and updating a product","example":"test title"},"description":{"type":"string","description":"product description","example":"test description"},"language":{"$ref":"#/components/schemas/Language"},"price":{"minimum":1,"type":"integer","description":"the price of product, required field when creating and updating a product","format":"int64","example":200},"priceCurrency":{"$ref":"#/components/schemas/PriceCurrency"},"creationDate":{"type":"string","description":"product creation date, when creating a product, it is filled in automatically","format":"date-time"},"modificationDate":{"type":"string","description":"product modification date, when updating and creating product, it is filled in automatically","format":"date-time"}},"description":"Product"}}}}